+++ components/layout.tsx
@@ -1,4 +1,4 @@
-import { Navbar } from './Navbar'
+import { Navbar } from './Navbar/Navbar'
 import Head from "next/head";
 
 export default function Layout({ children }) {
