+++ pages/dashboard/index.tsx
@@ -1,0 +1,11 @@
+import Link from 'next/link'
+import Layout from '../components/Layout'
+
+export default function IntakeDashboard() {
+    return (
+        // Layout wraps this component in index.tsx
+        <div> 
+            <h1>Intake Dashboard</h1>
+        </div>
+    )
+}
