--- pages/intake-dashboard.tsx
@@ -1,9 +1,0 @@
-import Link from 'next/link'
-import Layout from '../components/layout'
-
-export default function IntakeDashboard() {
-    return (
-        // <Layout>
-        <h1>Intake Dashboard</h1>
-    )
-}
